gpu: ion: Change secure heap allocation restrictions
Currently, allocations for secure memory will fail if there
is any oustanding mapping present on a secure heap. This is
too restrictive for allowing protection of individual buffers
instead of the entire heap. The correct solution is to allow
secured allocations even if there are oustanding mappings on
the heap. The heap protect call should fail if there are
oustanding mappings. Any detection of oustanding mappings
when the heap is secured should be considered a bug