gpu: ion: Don't require buffers to be 1M aligned
When generating the scatter gather table, older secure
clients cannot support 1MB aligned allocations. Change
the code flow to only generate the sg_table with 1MB
chunks if the buffer is 1MB aligned. Checks are present
elsewhere which will prevent the buffer from being
secured if the allocation is not 1MB aligned.