gpu: ion: fix race condition in checking virtual address
Currently, on an Ion cache operation, there is a short period of time
between checking the validity of the virtual address range of the
operation and the actual cache operation. During this time another
thread could twiddle the vm areas, resulting in an invalid cache
operation. Fix this by holding mmap_sem of the active_mm of the
current thread across the address range check and the cache operation.